requirements.txt
click==7.1.1
Pillow==7.0.0
toml==0.10.0
Jinja2==2.11.2
rectpack==0.2.1